requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-stubs==4.2.3
django-stubs-ext==4.2.2
django-treebeard==4.7
djangorestframework==3.14.0
mypy==1.4.1
mypy-extensions==1.0.0
Pillow==10.0.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
typing_extensions==4.7.1
